.requirements.txt
pip~=22.2.2
cryptography~=39.0.2
Jinja2~=3.0.3
setuptools~=58.1.0
requests~=2.28.2
pymongo~=3.11.4
Flask~=2.2.3
python-decouple==3.8